Correct improve projectile targeting calculations

+++ src/helpers/Clock.java
@@ -62,7 +62,7 @@ public class Clock {
     }
     
     
-    public static void changeMultiplier(int change) {
+    public static void changeMultiplier(float change) {
         
         if(multiplier + change < -1 || multiplier + change > 7) {
             // ???
